Centering of boardwindow now present
Upon initiating the game the window will be set to fullscreen if the option is selected, and the window will have the board positioned in the center of the screen. It will not wait until there is a resize event to center the board.

Will next tie in the resolution options to that selected in the main menu and have it saved across closing and reopening the application.